django/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bleach==5.0.1
Django==4.1
django-bootstrap-v5==1.0.11
django-js-asset==2.0.0
django-summernote==0.8.20.0
djangorestframework==3.13.1
mysqlclient==2.1.1
pytz==2022.2.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.2
webencodings==0.5.1
wincertstore==0.2
